requirements.txt
asgiref==3.8.1
Django==5.0.3
psycopg2-binary==2.9.9
sqlparse==0.4.4
gunicorn==21.2.0
boto3==1.34.84
django-storages==1.14.2
Pillow==10.3.0
djangorestframework==3.15.1
